+++ Assets/Scripts/Inventory System/ActiveInvetoriesHandler.cs
@@ -24,12 +24,12 @@ public class ActiveInvetoriesHandler : MonoBehaviour
         if (itemInventory == _playerInventory)
         {
             _playerInventory.ResetItemServerRpc(itemDisplayer.PreviousCell.Index);
-            _activeInventory.AddItemToDesiredSlotServerRpc(itemDisplayer.InventoryCell.Item.Id, itemDisplayer.InventoryCell.Count);
+            _activeInventory.AddItemToDesiredSlotServerRpc((ushort)itemDisplayer.InventoryCell.Item.Id, (ushort)itemDisplayer.InventoryCell.Count);
         }
         else
         {
             _activeInventory.ResetItemServerRpc(itemDisplayer.PreviousCell.Index);
-            _playerInventory.AddItemToDesiredSlotServerRpc(itemDisplayer.InventoryCell.Item.Id, itemDisplayer.InventoryCell.Count);
+            _playerInventory.AddItemToDesiredSlotServerRpc((ushort)itemDisplayer.InventoryCell.Item.Id, (ushort)itemDisplayer.InventoryCell.Count);
         }
         Destroy(itemDisplayer.gameObject);
     }
