+++ Assets/Scripts/Player Controller/PlayerResourcesGatherer.cs
@@ -132,7 +132,7 @@ public class PlayerResourcesGatherer : MonoBehaviour
     {
         var lootingItem = _objectsRayCaster.LootingItem;
         if(!lootingItem) return;
-        _inventoryHandler.CharacterInventory.AddItemToDesiredSlotServerRpc(lootingItem.ItemId.Value, lootingItem.Count.Value);
+        _inventoryHandler.CharacterInventory.AddItemToDesiredSlotServerRpc((ushort)lootingItem.ItemId.Value, (ushort)lootingItem.Count.Value);
         lootingItem.PickUpServerRpc();
     }
 }
